requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
mysql-connector-python==8.2.0
influxdb-client==1.38.0
pandas==2.1.3
numpy==1.24.3
avro-python3==1.11.2
fastavro==1.9.0
python-dotenv==1.0.0
schedule==1.2.0
confluent-kafka==2.3.0
pyarrow==14.0.1
sqlalchemy==2.0.23
apache-airflow==2.7.3
requests==2.31.0
pyyaml==6.0.1
